added auto increment ids

+++ src/database/schemas/customers/customers.sql
@@ -1,6 +1,6 @@
 --@block
 CREATE TABLE "Customers" (
-  "CustomerID" bigint PRIMARY KEY,
+  "CustomerID" SERIAL PRIMARY KEY,
   "Name" varchar(255),
   "Email" varchar(255) UNIQUE,
   "PhoneNumber" varchar(255)
